Sync

+++ graphics/src/Metrics.hpp
@@ -58,11 +58,11 @@ namespace jpl{
              */
             extern void windowPositionCallback(GLFWwindow* window, int x, int y);
 
-            inline float XPosToScreenCoords(unsigned int x){
+            inline float XPosToScreenCoords(float x){
                 float halfW = ((float)jpl::_graphics::_metrics::width)/2;
                 return (x-halfW)/halfW;
             }
-            inline float YPosToScreenCoords(unsigned int y){
+            inline float YPosToScreenCoords(float y){
                 float halfH = ((float)jpl::_graphics::_metrics::height)/2;
                 return (y-halfH)/halfH;
             }
